Modified snowdepth decoder to handle json payload
* Some values seem wonky. We need to dig into those a bit more before passing them on to subscribers.

+++ Masarin.IoT.Sensor/Messages/TelemetrySnowdepth.cs
@@ -2,9 +2,9 @@
 {
 	class TelemetrySnowdepth : IoTHubMessage
     {
-        public int Snowdepth { get; }
+        public double Snowdepth { get; }
 
-        public TelemetrySnowdepth(IoTHubMessageOrigin origin, string timestamp, int snowdepth) : base(origin, timestamp, "telemetry.Snowdepth")
+        public TelemetrySnowdepth(IoTHubMessageOrigin origin, string timestamp, double snowdepth) : base(origin, timestamp, "telemetry.Snowdepth")
         {
             Snowdepth = snowdepth;
         }
